Support decimal handicaps
Parse and accept decimal (tenths) handicap values instead of rounding to integers. Replaced parseInt with parseFloat when creating/updating players and added step="0.1" to the handicap inputs in AdminPage so users can enter fractional 18-hole handicaps. Files changed: src/LeagueContext.jsx, src/pages/AdminPage.jsx.

+++ src/LeagueContext.jsx
@@ -595,7 +595,7 @@ export function LeagueProvider({ children }) {
       ...newPlayer,
       id: maxId + 1,
       name: newPlayer.name.trim(),
-      handicap: parseInt(newPlayer.handicap) || 0,
+      handicap: parseFloat(newPlayer.handicap) || 0,
       weeksPlayed: 0,
       totalMoney: 0,
       weeklyMoney: {}
